+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changes:
 
 \-
 
+## v2023-09-20-1
+
+Changes:
+
+- Remove inline comments
+
 ## v2023-09-19-2
 
 Changes: